+++ src/UofIAxonius/UofIAxonius.psm1
@@ -1,11 +1,11 @@
-# Use environment variable for TDXSettings (explained in README.md)
+# Use environment variable for AxoniusSettings (explained in README.md)
 if ($env:AxoniusSettings) {
     $script:Settings=$env:AxoniusSettings | ConvertFrom-Json
 }
 else {
     Write-Error -Message "No environment variable AxoniusSettings found.
-    Set `$env:AxoniusSettings to a JSON-formatted string containing a BaseURI property, for example:
-    `$env:AxoniusSettings='{ ""BaseURI"": [""https://{axonius-URI}:{port}/api/""] }'"
+    Set `$ENV:AxoniusSettings to a JSON-formatted string containing a BaseURI property, for example:
+    `$ENV:AxoniusSettings='{ ""BaseURI"": [""https://{axonius-URI}:{port}/api/v2/""] }'"
 }
 
 $Script:Session = $NULL