+++ CHANGELOG.md
@@ -9,6 +9,32 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Removed
 
+## [0.0.2] - 2025-03-11
+
+### Added
+
+- New-AxoniusSession
+- Invoke-AxoniusRestCall
+- Add-AxoniusAssetLink
+- Add-AxoniusTags
+- Get-AxoniusAssetByID
+- Get-AxoniusAssetCount
+- Get-AxoniusAssetFields
+- Get-AxoniusAssets
+- Get-AxoniusQueries
+- Remove-AxoniusAssetLink
+- Remove-AxoiusTag
+- CONTRIBUTING.md
+- icon.png
+
+### Changed
+
+- Supported version and reporting vulnerabilities language in SECURITY.md
+- Standard module settings and link to icon in UofIAxonius.psd1
+- Base URI example needed a /v2/ in UofIAxonius.psm1
+- Reference to TDX replaced with Axonius in Tests
+
+
 ## [0.0.1] - 2025-03-05
 
 ### Added